Move demo to /demo/* route hierarchy with SSR support
- Replace query-param navigation (/?view=all, /?sub={id}) with path-based
  routes (/demo/all, /demo/subscription/{id}, /demo/tag/{id}, etc.)
- Split monolithic DemoPage into DemoLayoutContent, DemoRouter, DemoEntryList
  (mirrors AppLayoutContent + AppRouter pattern)
- Add Next.js route stubs with metadata for SEO (generateMetadata for dynamic routes)
- Make / redirect to /demo/all?entry=welcome (unauthenticated) or /all (authenticated)
- Demo pages accessible to both logged-in and logged-out users
