+++ src/app/demo/page.tsx
@@ -1,0 +1,9 @@
+/**
+ * /demo → redirect to /demo/all?entry=welcome
+ */
+
+import { redirect } from "next/navigation";
+
+export default function DemoIndexPage() {
+  redirect("/demo/all?entry=welcome");
+}
